refactor: address pre-PR review (extract pull machinery, reset retry parser, lock + doc fixes)
