+++ cli/cmd/start.go
@@ -10,7 +10,6 @@ import (
 	"path/filepath"
 	"strconv"
 	"strings"
-	"sync"
 	"time"
 
 	"github.com/Aureliolo/synthorg/cli/internal/compose"
@@ -336,78 +335,6 @@ func startDetached(ctx context.Context, info docker.Info, safeDir string, state 
 	return nil
 }
 
-// pullAllImages pulls all enabled images in a single unified LiveBox:
-// compose services (backend, web, postgres, nats) plus standalone images
-// (sandbox, sidecar, fine-tune) depending on configuration. Only enabled
-// services are pulled.
-//
-// Callers MUST pass a state whose ImageTag and VerifiedDigests reflect the
-// images to be pulled. During an update, disk config still holds the old
-// tag/digests until after the pull completes; reloading here would cause
-// standalone image pulls to use stale refs while compose-driven pulls use
-// the new refs written into compose.yml, leaving the install inconsistent.
-func pullAllImages(ctx context.Context, cmd *cobra.Command, info docker.Info, safeDir string, state config.State, out *ui.UI) (config.State, error) {
-	if stateHasRegistryOverrides(state) {
-		warnRegistryOverridesDisableVerification(cmd)
-	}
-	items := buildPullItems(state)
-	emitFineTuneSizeHint(state, out)
-	return state, runPullBatch(ctx, info, safeDir, items, out)
-}
-
-// pullItem describes one image to pull. compose=true uses
-// `docker compose pull <name>`; compose=false uses `docker pull <ref>`
-// with retry/backoff.
-type pullItem struct {
-	name    string
-	compose bool
-	ref     string
-}
-
-// buildPullItems enumerates every image the start path must pull: the
-// enabled compose services plus the standalone (sandbox / sidecar /
-// fine-tune) images that compose does not own.
-//
-// When the operator has overridden any of the registry / image-tag
-// tunables (registry_host, image_repo_prefix, dhi_registry,
-// postgres_image_tag, nats_image_tag), state.VerifiedDigests is bound
-// to the DEFAULT-registry images and would pin the standalone pulls
-// to stale digests that do not exist on the override registry. Drop
-// the digest in that case so the pull resolves the tag on the
-// override registry instead of failing on a stale @sha256 reference.
-func buildPullItems(state config.State) []pullItem {
-	var items []pullItem
-	for _, svc := range composeServiceNames(state) {
-		items = append(items, pullItem{name: svc, compose: true})
-	}
-	useDigests := !stateHasRegistryOverrides(state)
-	pickDigest := func(name string) string {
-		if !useDigests {
-			return ""
-		}
-		return state.VerifiedDigests[name]
-	}
-	if state.Sandbox {
-		items = append(items, pullItem{
-			name: "sandbox",
-			ref:  verify.FormatImageRef("sandbox", state.ImageTag, pickDigest("sandbox")),
-		})
-		items = append(items, pullItem{
-			name: "sidecar",
-			ref:  verify.FormatImageRef("sidecar", state.ImageTag, pickDigest("sidecar")),
-		})
-	}
-	if state.FineTuning {
-		variant := state.FineTuneVariantOrDefault()
-		svc := verify.FineTuneServiceName(variant)
-		items = append(items, pullItem{
-			name: svc,
-			ref:  verify.FormatImageRef(svc, state.ImageTag, pickDigest(svc)),
-		})
-	}
-	return items
-}
-
 // registryOverrideEnvVars lists every env var that, if set, overrides
 // a registry / image-tag tunable for the current invocation. Mirrors
 // the env precedence inputs ResolveTunables uses; checking these
@@ -461,100 +388,6 @@ func warnRegistryOverridesDisableVerification(cmd *cobra.Command) {
 	)
 }
 
-// emitFineTuneSizeHint warns the user about the fine-tune image size
-// BEFORE the pull box renders so they understand why their terminal is
-// about to pause. Emitting it after (the old behaviour) was a logic
-// error: by the time the warning appeared, the wait had already
-// completed.
-//
-// Uses HintNextStep (not HintGuidance): in the default "auto" hints mode
-// HintGuidance is suppressed entirely, so the warning that explains a
-// 10-to-30-minute stall was invisible to exactly the operators who needed
-// it. The GPU figures distinguish download size from the larger on-disk
-// footprint and give an honest duration: the GPU image bundles ~2.5 GB of
-// CUDA wheels inside the torch wheel, so a first pull realistically takes
-// far longer than "a few minutes".
-func emitFineTuneSizeHint(state config.State, out *ui.UI) {
-	if !state.FineTuning {
-		return
-	}
-	sizeHint := "~4 GB download (~7 GB on disk); first pull commonly takes 10-30 minutes"
-	if state.FineTuneVariantOrDefault() == config.FineTuneVariantCPU {
-		sizeHint = "~1.7 GB; first pull usually takes a few minutes"
-	}
-	out.HintNextStep(fmt.Sprintf(
-		"Fine-tune image is %s on typical connections.",
-		sizeHint,
-	))
-}
-
-// runPullBatch fans out a pull goroutine per item and renders progress
-// in a single LiveBox. Each line shows live pull progress (downloaded bytes
-// and layer counts where Docker emits them, plus elapsed time) instead of a
-// static spinner, so a multi-minute image pull shows activity. Returns the
-// joined error covering every failed pull (nil when every pull succeeds).
-func runPullBatch(ctx context.Context, info docker.Info, safeDir string, items []pullItem, out *ui.UI) error {
-	labels := make([]string, len(items))
-	for i, item := range items {
-		labels[i] = item.name
-	}
-	lb := out.NewLiveBoxWithProgress("Pull Images", labels)
-	defer lb.Finish()
-
-	var (
-		mu      sync.Mutex
-		pullErr error
-		wg      sync.WaitGroup
-	)
-	for i, item := range items {
-		wg.Add(1)
-		go func(idx int, it pullItem) {
-			defer wg.Done()
-			// Each goroutine owns its parser; the LiveBox line updates only
-			// when the parsed state advances, avoiding redundant redraws.
-			pp := &docker.PullProgress{}
-			onLine := func(line string) {
-				if pp.Observe(line) {
-					lb.UpdateProgress(idx, pp.Render())
-				}
-			}
-			err := pullOneItem(ctx, info, safeDir, it, onLine)
-			if err != nil {
-				lb.UpdateLine(idx, ui.IconError)
-				mu.Lock()
-				pullErr = errors.Join(pullErr, fmt.Errorf("pulling %s: %w", it.name, err))
-				mu.Unlock()
-				return
-			}
-			lb.UpdateLine(idx, ui.IconSuccess)
-		}(i, item)
-	}
-	wg.Wait()
-	return pullErr
-}
-
-// pullOneItem dispatches to the right puller for the item kind: compose
-// services go through docker-compose's own pull (so it picks up the
-// image override from compose.yml); standalone images use the retrying
-// dockerPullWithRetry. onLine receives each line of pull output for live
-// progress rendering.
-func pullOneItem(ctx context.Context, info docker.Info, safeDir string, it pullItem, onLine func(string)) error {
-	if it.compose {
-		return composeRunStreaming(ctx, info, safeDir, onLine, "pull", it.name)
-	}
-	tun := GetGlobalOpts(ctx).Tunables
-	return dockerPullWithRetry(ctx, info, it.ref, tun.ImagePullAttempts, tun.ImagePullRetryDelay, onLine)
-}
-
-// maxPullBackoff caps the exponential-backoff delay between image-pull
-// retries.  Guards against int64 overflow when operators set
-// ImagePullAttempts to a high value: “baseDelay << (attempt - 1)“
-// with a 2-second base and >=62 attempts would overflow time.Duration
-// (int64 nanoseconds) and yield a negative delay that time.After
-// resolves immediately, effectively disabling the backoff.  Saturating
-// at 5 minutes keeps the retry schedule bounded and predictable.
-const maxPullBackoff = 5 * time.Minute
-
 // Health-check polling cadence shared by both start paths
 // (startDetached and pullStartAndWait).
 //
@@ -576,65 +409,6 @@ const (
 	healthInitialDelay = 5 * time.Second
 	dhiVerifyTimeout   = 120 * time.Second
 )
-
-// dockerPullWithRetry pulls an image with retries for transient failures.
-// The caller supplies attempts (> 0) and baseDelay (exponential backoff
-// seed) so the values flow from the resolved
-// config.Tunables rather than being pinned by package-level constants.
-func dockerPullWithRetry(
-	ctx context.Context,
-	info docker.Info,
-	imageRef string,
-	attempts int,
-	baseDelay time.Duration,
-	onLine func(string),
-) error {
-	var lastErr error
-	for attempt := 1; attempt <= attempts; attempt++ {
-		err := dockerRunStreaming(ctx, info, onLine, "pull", imageRef)
-		if err == nil {
-			return nil
-		}
-		lastErr = err
-		if attempt == attempts || ctx.Err() != nil {
-			break
-		}
-		backoff := computePullBackoff(baseDelay, attempt)
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case <-time.After(backoff):
-		}
-	}
-	return lastErr
-}
-
-// computePullBackoff returns “baseDelay << (attempt - 1)“ saturated
-// at maxPullBackoff and guarded against int64 overflow.  Attempt is
-// 1-indexed.
-func computePullBackoff(baseDelay time.Duration, attempt int) time.Duration {
-	// Any baseDelay <= 0 from a misconfigured caller collapses to the
-	// ceiling immediately -- better than a zero-wait tight retry loop.
-	if baseDelay <= 0 {
-		return maxPullBackoff
-	}
-	// Compute the shift amount that would exceed the ceiling so we
-	// can bail out before int64 overflow.  The ceiling is checked
-	// against every candidate backoff.
-	shift := max(attempt-1, 0)
-	// math.MaxInt64 / baseDelay gives the largest multiplier that
-	// stays in range; log2 of that caps the safe shift.
-	for range shift {
-		if baseDelay > maxPullBackoff/2 {
-			return maxPullBackoff
-		}
-		baseDelay *= 2
-	}
-	if baseDelay > maxPullBackoff {
-		return maxPullBackoff
-	}
-	return baseDelay
-}
 
 // pullStartAndWait pulls images, starts containers, and waits for health.
 func pullStartAndWait(ctx context.Context, cmd *cobra.Command, info docker.Info, safeDir string, state config.State, out, errOut *ui.UI) error {
@@ -911,119 +685,6 @@ func composeRunQuiet(ctx context.Context, info docker.Info, dir string, args ...
 		return err
 	}
 	return nil
-}
-
-// maxPullTailLines bounds the rolling buffer of recent output lines kept
-// for the error message of a failed streaming pull, so a verbose failure
-// does not balloon the wrapped error.
-const maxPullTailLines = 20
-
-// lineSplitter is an io.Writer that reassembles newline-delimited output
-// from a child process, forwarding each complete line to onLine and
-// retaining the last maxPullTailLines for error context. It is safe to use
-// as both Stdout and Stderr of a single command: writes are serialised
-// under mu so interleaved stdout/stderr chunks never corrupt a line mid-way.
-type lineSplitter struct {
-	mu     sync.Mutex
-	buf    []byte
-	onLine func(string)
-	tail   []string
-}
-
-func (l *lineSplitter) Write(p []byte) (int, error) {
-	l.mu.Lock()
-	defer l.mu.Unlock()
-	l.buf = append(l.buf, p...)
-	for {
-		i := bytes.IndexByte(l.buf, '\n')
-		if i < 0 {
-			break
-		}
-		line := string(l.buf[:i])
-		l.buf = l.buf[i+1:]
-		l.handleLocked(line)
-	}
-	return len(p), nil
-}
-
-// handleLocked forwards one complete line and folds it into the tail
-// buffer. Must be called with l.mu held.
-func (l *lineSplitter) handleLocked(line string) {
-	line = strings.TrimRight(line, "\r")
-	if l.onLine != nil && line != "" {
-		l.onLine(line)
-	}
-	l.tail = append(l.tail, line)
-	if len(l.tail) > maxPullTailLines {
-		l.tail = l.tail[len(l.tail)-maxPullTailLines:]
-	}
-}
-
-// flush emits any trailing partial line the process wrote without a final
-// newline (Docker's last status line sometimes lacks one).
-func (l *lineSplitter) flush() {
-	l.mu.Lock()
-	defer l.mu.Unlock()
-	if len(l.buf) > 0 {
-		l.handleLocked(string(l.buf))
-		l.buf = nil
-	}
-}
-
-// tailString returns the sanitized recent output for an error message.
-func (l *lineSplitter) tailString() string {
-	l.mu.Lock()
-	defer l.mu.Unlock()
-	return sanitizeCLIOutput(strings.Join(l.tail, "\n"))
-}
-
-// composeRunStreaming runs a docker compose command, forwarding each line of
-// combined output to onLine for live progress rendering. It is the streaming
-// counterpart of composeRunQuiet: on error the recent output is folded into
-// the error message, but the verbose progress never reaches the terminal
-// directly (the LiveBox owns the display).
-func composeRunStreaming(ctx context.Context, info docker.Info, dir string, onLine func(string), args ...string) error {
-	fullArgs := make([]string, 0, len(info.ComposeCmd)-1+len(args))
-	fullArgs = append(fullArgs, info.ComposeCmd[1:]...)
-	fullArgs = append(fullArgs, args...)
-
-	split := &lineSplitter{onLine: onLine}
-	c := exec.CommandContext(ctx, info.ComposeCmd[0], fullArgs...) //nolint:gosec // G204: compose binary is CLI-detected (info.ComposeCmd), args internally assembled, never attacker-controlled
-	c.Dir = dir
-	c.Stdout = split
-	c.Stderr = split
-	err := c.Run()
-	split.flush()
-	return streamingError(err, split)
-}
-
-// dockerRunStreaming runs a docker command, forwarding each line of combined
-// output to onLine for live progress rendering. Streaming counterpart of
-// dockerRunQuiet.
-func dockerRunStreaming(ctx context.Context, info docker.Info, onLine func(string), args ...string) error {
-	dockerBin := info.DockerPath
-	if dockerBin == "" {
-		dockerBin = "docker"
-	}
-	split := &lineSplitter{onLine: onLine}
-	c := exec.CommandContext(ctx, dockerBin, args...) //nolint:gosec // G204: dockerBin is the resolved docker binary (info.DockerPath), args internally assembled, never attacker-controlled
-	c.Stdout = split
-	c.Stderr = split
-	err := c.Run()
-	split.flush()
-	return streamingError(err, split)
-}
-
-// streamingError wraps a streaming command's exit error with its recent
-// sanitized output, mirroring the error shape of the *Quiet helpers.
-func streamingError(err error, split *lineSplitter) error {
-	if err == nil {
-		return nil
-	}
-	if output := split.tailString(); output != "" {
-		return fmt.Errorf("%w: %s", err, output)
-	}
-	return err
 }
 
 // sanitizeCLIOutput strips control characters from external CLI output
